Github Desktop was bugged, needed to reinstall so couldnt push for a while so bigger push now (check description)
Added mainwindowVM for switching pages from category to items to itemdetails. added functionality to get the total amount of items that exist in a specific category, improved many visuals to make the application fall in line with the Runescape 3 graphical layout. Itemdetailpage adapts to current item (not image tho), items in app vs in total existing is shown throught the app